Improved unittests
- Part of unit testing is passing improper arguments to the method
being tested and ensuring the method handles invalid input
appropriately. This unit test showed that the code did not originally
handle those conditions properly, so the code had to be changed. This
is a good example of how unites improve your code and force one to
think of things that had not originally considered.